feat: enable ccache and lld linker in CMake, implement LogicComparator::extractVariables() method
