+++ README.md
@@ -332,3 +332,11 @@ Subcommands:
       - 当前 token 不是一个显式逻辑运算符 (`and`, `or`, `xor`, `not`) 或左括号 `(`。
       - 下一个 token 不是一个显式逻辑运算符或右括号 `)`。
   - **重组与清理**：将处理后的 token 列表用单个空格连接起来，并进行最终的空格清理（去除多余空格，修剪首尾空格），得到最终的预处理表达式。
+- 在 `CMakeLists.txt` 中，强制启用了 lld 链接器和 ccache，以提升链接和编译速度。
+- 为了保持 `main.cpp` 的整洁，将 `LibFile()` 函数的相关操作剥离出来，并放入 `LibFileOperations.cpp` 文件中。
+- 实现了 `LogicComparator::extractVariables()` 方法，用于从两个表达式中提取并验证变量名，确保它们一致。
+  - 该方法使用正则表达式初步提取所有可能的标识符。
+  - 对每个提取出的标识符，先用 `isIdentifier` 函数（检查大写开头等）进行验证。
+  - 通过 `isIdentifier` 验证后，仍然将其转换为小写并与 `keywords` 集合比较，以过滤掉可能的关键词（这是一个保守但安全的操作）。
+  - 只有通过这两步验证的标识符才被认为是有效变量，并以原始大小写形式添加到各自的 `std::set` 中。
+  - 最后比较两个 `std::set` 是否相等，如果不等则报告差异并返回 `false`，如果相等则将变量列表（保留原始大小写）排序后存入 `sorted_vars` 并返回 `true`。
