+++ include/LogicComparator.hpp
@@ -47,6 +47,24 @@ public:
   // Preprocessing function
   std::string preprocessExpression(const std::string &input_expr);
 
+  // Helper to extract unique sorted variables from TWO expressions
+  bool extractVariables(const std::string &expr1, const std::string &expr2,
+                        std::vector<std::string> &sorted_vars);
+
+  /**
+   * @brief Compares two preprocessed logic expression strings for equivalence.
+   *        Internal helper function.
+   *
+   * @tparam T The numeric type used by ExprTk (e.g., double, float).
+   * @param ref_expression_processed The preprocessed reference logic expression.
+   * @param comp_expression_processed The preprocessed comparison logic expression.
+   * @param result Reference to a PinComparisonResult object to store detailed results.
+   */
+  template <typename T>
+  void compareSingleExpressionPair(const std::string &ref_expression_processed,
+                                   const std::string &comp_expression_processed,
+                                   PinComparisonResult &result); // Pass result struct
+
   /**
    * @brief Compares logic for all output pins defined in the input maps.
    *
@@ -69,24 +87,6 @@ private:
   std::map<std::string, std::string> ref_outpin_map_;
   std::map<std::string, std::string> comp_outpin_map_;
   std::string cell_name_;
-
-  // Helper to extract unique sorted variables from TWO expressions
-  bool extractVariables(const std::string &expr1, const std::string &expr2,
-                        std::vector<std::string> &sorted_vars);
-
-  /**
-   * @brief Compares two preprocessed logic expression strings for equivalence.
-   *        Internal helper function.
-   *
-   * @tparam T The numeric type used by ExprTk (e.g., double, float).
-   * @param ref_expression_processed The preprocessed reference logic expression.
-   * @param comp_expression_processed The preprocessed comparison logic expression.
-   * @param result Reference to a PinComparisonResult object to store detailed results.
-   */
-  template <typename T>
-  void compareSingleExpressionPair(const std::string &ref_expression_processed,
-                                   const std::string &comp_expression_processed,
-                                   PinComparisonResult &result); // Pass result struct
 };
 
 #endif // LOGIC_COMPARATOR_HPP
